% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{convert_igraph.gng}
\alias{convert_igraph}
\alias{convert_igraph.gng}
\title{convert_igraph}
\format{\preformatted{ NULL
}}
\usage{
convert_igraph(gng)
}
\description{
Converts to igraph (O(n) method, writing intermediately to disk)
}
\examples{
convert_igraph(gng)
}
\keyword{datasets}

