% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{convertToGraph.gng}
\alias{convertToGraph}
\alias{convertToGraph.gng}
\title{convertToGraph}
\format{\preformatted{ NULL
}}
\usage{
convertToGraph(gng)
}
\description{
Converts to igraph (O(n) method, writing intermediately to disk)
}
\examples{
convertToGraph(gng)
}
\keyword{datasets}

