% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{errorStatistics.gng}
\alias{errorStatistics}
\alias{errorStatistics.gng}
\title{errorStatistics}
\usage{
errorStatistics(gng)

errorStatistics(gng)
}
\description{
Gets vector with errors for every second of execution

Gets vector with errors for every second of execution
}
\examples{
errorStatistics(gng)
errorStatistics(gng)
}

