% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{insertExamples.gng}
\alias{insertExamples}
\alias{insertExamples.gng}
\title{insertExamples}
\usage{
insertExamples(gng, M, L=c())
}
\arguments{
\item{examples}{Matrix with examples}
}
\description{
Insert (inefficiently) examples to algorithm dataset. For
efficient insertion use gng$inser_examples
}
\examples{
#Add preset examples
M = generateExamples(preset=gng.preset.sphere)
insertExamples(gng, M)
}

