% Generated by roxygen2 (4.0.2): do not edit by hand
\name{OptimizedGNG}
\alias{OptimizedGNG}
\title{Constructor of Optimized GrowingNeuralGas object.}
\usage{
OptimizedGNG(x = NULL, labels = c(), beta = 0.99, alpha = 0.5,
  max.nodes = 1000, eps.n = 6e-04, eps.w = 0.05, max.edge.age = 200,
  training = gng.train.offline(), lambda = 200, verbosity = 0,
  value.range = c(0, 1))
}
\arguments{
\item{beta}{coefficient. Decrease the error variables of all node nodes by this fraction. Forgetting rate. Default 0.99}

\item{alpha}{Alpha coefficient. Decrease the error variables of the nodes neighboring to the newly inserted node by this fraction. Default 0.5}

\item{max.nodes}{Maximum number of nodes (after reaching this size it will continue running, but won't add new nodes)}

\item{eps.n}{Default 0.0006. How strongly adapt neighbour node}

\item{eps.w}{Default 0.05. How strongly adapt winning node}

\item{training}{Can be either gng.train.offline(max.iter, min.improvement), or gng.train.online()}

\item{lambda}{Every lambda iteration is added new vertex. Default 200}

\item{value.range}{Default [0,1]. All example features should be in this range}
}
\description{
Construct simplified and optimized GNG object. Can be used to train offline, or online. Data dimensionality shouldn't be too big, if
it is consider using dimensionality reduction (for instance PCA). Cannot use utility.
}
\examples{
# Train online optimizedGNG. All values in this dataset are in the range (-4.3, 4.3)
data(wine, package="rattle")
gng <- OptimizedGNG(training = gng.train.online(), value.range=c(min(scale(wine[-1])),max(scale(wine[-1]))), max.nodes=20)
insertExamples(gng, scale(wine[-1]))
run(gng)
Sys.sleep(10)
pause(gng)
}

