% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{save.gng}
\alias{save.gng}
\title{save.gng}
\usage{
save.gng(gng)
}
\arguments{
\item{filename}{Dump destination}
}
\description{
Writes model to a disk space efficient binary format.
}
\details{
Save model to binary format
}
\examples{
save.gng(gng, 'graph.bin')
}

