\docType{methods}
\name{error_statistics.gng}
\alias{error_statistics}
\alias{error_statistics.gng}
\title{error_statistics}
\format{\preformatted{ NULL
}}
\usage{
error_statistics(gng)
}
\description{
Gets vector with errors for every second of execution
}
\examples{
error_statistics(gng)
}
\keyword{datasets}

